src/agentres/requirements.txt
# Core dependencies
python-dotenv>=1.0.0
requests
aiohttp
beautifulsoup4
selenium
playwright
pytest
pytest-asyncio
pytest-cov
pytest-mock
fastlogging
sqlmodel
prometheus_client

# Search and scraping
duckduckgo-search
google-api-python-client
tavily-python

# Browser automation
undetected-chromedriver
webdriver-manager

# Text processing and NLP
nltk
spacy
transformers
sentence-transformers
torch
tensorflow

# Utilities
tqdm
colorama
rich
pyyaml
python-json-logger
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-requests
opentelemetry-instrumentation-aiohttp-client

# Development tools
black
flake8
mypy
isort
pre-commit

# Documentation
sphinx
sphinx-rtd-theme
sphinx-autodoc-typehints

# Testing and monitoring
pytest-benchmark
pytest-xdist
pytest-timeout
pytest-env
pytest-sugar
pytest-html
pytest-metadata
pytest-randomly
pytest-repeat
pytest-rerunfailures
pytest-selenium
pytest-playwright

# Database dependencies
qdrant-client>=1.6.0
numpy>=1.24.0
sqlite3
openai>=1.0.0